requirements/base.txt
-r openstack.txt
-r hermes.txt
# Core
Django>=2.2,<3.0  # Updated to satisfy channels requirement
djangorestframework>=3.12.0,<4.0
channels==3.0.4

# Django Extensions
dj-choices==0.11.0
django-cors-headers>=3.10.0
django-filter>=2.0.0
django-import-export>=1.2.0
django-money>=0.12.3
django-mptt>=0.11.0
django-reversion>=3.0.5
django-rq>=2.0
django-sitetree==1.17.2
django-taggit>=1.3.0
django-taggit-serializer>=0.1.7
django-threadlocals>=0.10
drf-nested-routers>=0.93.4


factory-boy==2.11.1
mysqlclient==1.3.13
netaddr==0.7.18
openpyxl==2.4.0
py-moneyed==1.2
python-dateutil==2.4.2
pytz==2015.4
redis==3.2.1
requests-oauthlib==1.3.0
requests==2.20.0
rq==1.0
six>=1.9.0
sqlparse==0.2.3
tablib==0.11.5
typing==3.6.6
